+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed enum with comments parsing (see #188).
+- Fixed a bug which would cause pressing enter after the end of a block comment to add an unwanted space to the indentation.
 
 ## [3.0.10]